update to deserialize while and compound statement constructs

+++ pkg/codegen/backend/src/deserialize/Block.cc
@@ -6,7 +6,7 @@
 #include "proto/program.pb.h"
 
 Block::Block(const Pasc::Block &blk) {
-  for (size_t i = 0; i < blk.vardecl_size(); i++) {
+  for (int i = 0; i < blk.vardecl_size(); i++) {
     auto &varDecl = blk.vardecl(i);
     VarDeclrs.push_back(
         std::make_unique<VariableDeclaration>(varDecl));
@@ -16,7 +16,7 @@ Block::Block(const Pasc::Block &blk) {
     callables.push_back(deserializeCallable(blk.callables(i)));
   }
 
-  for (size_t i = 0; i < blk.stmts_size(); i++) {
+  for (int i = 0; i < blk.stmts_size(); i++) {
     Stmts.push_back(deserializeStmt(blk.stmts(i)));
   }
 }
